+++ x/evm/types/interfaces.go
@@ -4,6 +4,7 @@ package types
 
 import (
 	context "context"
+	"math/big"
 
 	"cosmossdk.io/core/address"
 	"cosmossdk.io/math"
